taskmate-api/python_llm_service/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
